chore: split type and composable
