+++ frontend/app/src/modules/history/management/forms/EthDepositEventForm.spec.ts
@@ -1,5 +1,5 @@
 import type { AssetMap } from '@/types/asset';
-import type { EthDepositEvent } from '@/types/history/events';
+import type { EthDepositEvent } from '@/types/history/events/schemas';
 import { useAssetInfoApi } from '@/composables/api/assets/info';
 import { useHistoryEvents } from '@/composables/history/events';
 import EthDepositEventForm from '@/modules/history/management/forms/EthDepositEventForm.vue';
