+++ frontend/app/src/modules/history/management/forms/EthDepositEventForm.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { StandaloneEventData } from '@/modules/history/management/forms/form-types';
-import type { EthDepositEvent, NewEthDepositEventPayload } from '@/types/history/events';
+import type { EthDepositEvent, NewEthDepositEventPayload } from '@/types/history/events/schemas';
 import AmountInput from '@/components/inputs/AmountInput.vue';
 import AutoCompleteWithSearchSync from '@/components/inputs/AutoCompleteWithSearchSync.vue';
 import JsonInput from '@/components/inputs/JsonInput.vue';
