+++ frontend/app/src/modules/history/management/forms/EvmEventForm.spec.ts
@@ -1,5 +1,5 @@
 import type { AssetMap } from '@/types/asset';
-import type { EvmHistoryEvent } from '@/types/history/events';
+import type { EvmHistoryEvent } from '@/types/history/events/schemas';
 import type { TradeLocationData } from '@/types/history/trade/location';
 import { useAssetInfoApi } from '@/composables/api/assets/info';
 import { useHistoryEvents } from '@/composables/history/events';
