+++ frontend/app/src/modules/history/management/forms/evm-swap-event-form.ts
@@ -1,3 +1,3 @@
-import type { AddEvmSwapEventPayload } from '@/types/history/events';
+import type { AddEvmSwapEventPayload } from '@/types/history/events/schemas';
 
 export type EvmSwapFormData = Required<AddEvmSwapEventPayload>;
